# autotest-platform/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Avtolider API
  description: Uzbekistan UBDD Driving License Exam Preparation Platform
  version: 1.0.0
  contact:
    name: Avtolider Team
servers:
  - url: http://localhost:5000/api/v1
    description: Development
  - url: https://api.avtolider.uz/api/v1
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LocalizedText:
      type: object
      required: [uz, uzLatin, ru]
      properties:
        uz:
          type: string
          description: Uzbek Cyrillic
        uzLatin:
          type: string
          description: Uzbek Latin
        ru:
          type: string
          description: Russian
    PaginationMeta:
      type: object
      required: [page, pageSize, totalCount, totalPages]
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        totalCount:
          type: integer
        totalPages:
          type: integer
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
    ApiResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        data: {}
        error:
          $ref: '#/components/schemas/ApiError'

    # ─── Auth ───────────────────────────────────
...
    AnswerOptionDto:
      type: object
      required: [id, text]
      properties:
        id:
          type: string
          format: uuid
        text:
          $ref: '#/components/schemas/LocalizedText'
        imageUrl:
          type: string
          nullable: true
    QuestionDto:
      type: object
      required: [id, text, difficulty, categoryId, categoryName, ticketNumber, answerOptions]
      properties:
        id:
          type: string
          format: uuid
        text:
          $ref: '#/components/schemas/LocalizedText'
        imageUrl:
          type: string
          nullable: true
        difficulty:
          type: integer
          enum: [1, 2, 3]
        categoryId:
          type: string
          format: uuid
        categoryName:
          $ref: '#/components/schemas/LocalizedText'
        ticketNumber:
          type: integer
        answerOptions:
          type: array
          items:
            $ref: '#/components/schemas/AnswerOptionDto'

    # ─── Exams ──────────────────────────────────
...
paths:
...
  /admin/questions:
    get:
      tags: [Admin Questions]
      operationId: adminGetQuestions
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: string
            format: uuid
        - name: isActive
          in: query
          schema:
            type: boolean
        - name: difficulty
          in: query
          schema:
            type: integer
            enum: [1, 2, 3]
        - name: ticketNumber
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Paginated admin questions
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          items:
                            type: array
                            items:
                              $ref: '#/components/schemas/QuestionDto'
                          meta:
                            $ref: '#/components/schemas/PaginationMeta'
    post:
      tags: [Admin Questions]
      operationId: adminCreateQuestion
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [categoryId, textUz, textUzLatin, textRu, difficulty, ticketNumber, licenseCategory]
              properties:
                categoryId:
                  type: string
                  format: uuid
                textUz:
                  type: string
                  description: Uzbek Cyrillic
                textUzLatin:
                  type: string
                  description: Uzbek Latin
                textRu:
                  type: string
                  description: Russian
                explanationUz:
                  type: string
                explanationUzLatin:
                  type: string
                explanationRu:
                  type: string
                difficulty:
                  type: integer
                  enum: [1, 2, 3]
                ticketNumber:
                  type: integer
                licenseCategory:
                  type: string
                  enum: [AB, CD, BOTH]
                isActive:
                  type: boolean
                  default: true
                questionImage:
                  type: string
                  format: binary
                options[0].textUz:
                  type: string
                options[0].textUzLatin:
                  type: string
                options[0].textRu:
                  type: string
                options[0].isCorrect:
                  type: boolean
                options[0].image:
                  type: string
                  format: binary
                options[1].textUz:
                  type: string
                options[1].textUzLatin:
                  type: string
                options[1].textRu:
                  type: string
                options[1].isCorrect:
                  type: boolean
                options[1].image:
                  type: string
                  format: binary
                options[2].textUz:
                  type: string
                options[2].textUzLatin:
                  type: string
                options[2].textRu:
                  type: string
                options[2].isCorrect:
                  type: boolean
                options[2].image:
                  type: string
                  format: binary
                options[3].textUz:
                  type: string
                options[3].textUzLatin:
                  type: string
                options[3].textRu:
                  type: string
                options[3].isCorrect:
                  type: boolean
                options[3].image:
                  type: string
                  format: binary
                options[4].textUz:
                  type: string
                options[4].textUzLatin:
                  type: string
                options[4].textRu:
                  type: string
                options[4].isCorrect:
                  type: boolean
                options[4].image:
                  type: string
                  format: binary
                options[5].textUz:
                  type: string
                options[5].textUzLatin:
                  type: string
                options[5].textRu:
                  type: string
                options[5].isCorrect:
                  type: boolean
                options[5].image:
                  type: string
                  format: binary
      responses:
        '201':
          description: Question created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/QuestionDto'
  /admin/questions/{id}:
    put:
      tags: [Admin Questions]
      operationId: adminUpdateQuestion
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                categoryId:
                  type: string
                  format: uuid
                textUz:
                  type: string
                  description: Uzbek Cyrillic
                textUzLatin:
                  type: string
                  description: Uzbek Latin
                textRu:
                  type: string
                  description: Russian
                explanationUz:
                  type: string
                explanationUzLatin:
                  type: string
                explanationRu:
                  type: string
                difficulty:
                  type: integer
                  enum: [1, 2, 3]
                ticketNumber:
                  type: integer
                licenseCategory:
                  type: string
                  enum: [AB, CD, BOTH]
                isActive:
                  type: boolean
                keepQuestionImage:
                  type: boolean
                questionImage:
                  type: string
                  format: binary
                keepOptionImages[0]:
                  type: boolean
                keepOptionImages[1]:
                  type: boolean
                keepOptionImages[2]:
                  type: boolean
                keepOptionImages[3]:
                  type: boolean
                keepOptionImages[4]:
                  type: boolean
                keepOptionImages[5]:
                  type: boolean
                options[0].textUz:
                  type: string
                options[0].textUzLatin:
                  type: string
                options[0].textRu:
                  type: string
                options[0].isCorrect:
                  type: boolean
                options[0].image:
                  type: string
                  format: binary
                options[1].textUz:
                  type: string
                options[1].textUzLatin:
                  type: string
                options[1].textRu:
                  type: string
                options[1].isCorrect:
                  type: boolean
                options[1].image:
                  type: string
                  format: binary
                options[2].textUz:
                  type: string
                options[2].textUzLatin:
                  type: string
                options[2].textRu:
                  type: string
                options[2].isCorrect:
                  type: boolean
                options[2].image:
                  type: string
                  format: binary
                options[3].textUz:
                  type: string
                options[3].textUzLatin:
                  type: string
                options[3].textRu:
                  type: string
                options[3].isCorrect:
                  type: boolean
                options[3].image:
                  type: string
                  format: binary
                options[4].textUz:
                  type: string
                options[4].textUzLatin:
                  type: string
                options[4].textRu:
                  type: string
                options[4].isCorrect:
                  type: boolean
                options[4].image:
                  type: string
                  format: binary
                options[5].textUz:
                  type: string
                options[5].textUzLatin:
                  type: string
                options[5].textRu:
                  type: string
                options[5].isCorrect:
                  type: boolean
                options[5].image:
                  type: string
                  format: binary
      responses:
        '200':
          description: Question updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/QuestionDto'
    delete:
      tags: [Admin Questions]
      operationId: adminSoftDeleteQuestion
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Question soft-deleted
  /admin/questions/{id}/permanent:
    delete:
      tags: [Admin Questions]
      operationId: adminPermanentDeleteQuestion
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: confirm
          in: query
          required: true
          schema:
            type: boolean
      responses:
        '204':
          description: Question permanently deleted
  /admin/questions/{id}/status:
    patch:
      tags: [Admin Questions]
      operationId: adminToggleQuestionStatus
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [isActive]
              properties:
                isActive:
                  type: boolean
      responses:
        '200':
          description: Status toggled
  /admin/questions/bulk-status:
    patch:
      tags: [Admin Questions]
      operationId: adminBulkToggleStatus
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [questionIds, isActive]
              properties:
                questionIds:
                  type: array
                  items:
                    type: string
                    format: uuid
                isActive:
                  type: boolean
      responses:
        '200':
          description: Bulk status updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          updatedCount:
                            type: integer
...
  /admin/questions/export-template:
    get:
      tags: [Admin Questions]
      operationId: adminExportTemplate
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Excel template
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary

  # ═══════════ ADMIN — DASHBOARD ═══════════
...